relocate random_internal
