+++ src/random/CMakeLists.txt
@@ -1,14 +1,9 @@
 target_sources_local(cytnx
     PRIVATE
-    random_internal_interface.hpp
-    random_internal_interface.cpp
+    #random_internal_interface.hpp
+    #random_internal_interface.cpp
     normal_.cpp
     uniform_.cpp
     normal.cpp
     uniform.cpp
 )
-
-add_subdirectory(random_internal_cpu)
-if(USE_CUDA)
-  add_subdirectory(random_internal_gpu)
-endif()
